# beryl-core-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beryl Core API",
    "description": "Central API Gateway and orchestration layer for Béryl ecosystem",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/aoq/signals": {
      "post": {
        "tags": [
          "AOQ"
        ],
        "summary": "Create Signal",
        "operationId": "create_signal_api_v1_aoq_signals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aoq/decision": {
      "post": {
        "tags": [
          "AOQ"
        ],
        "summary": "Create Decision",
        "operationId": "create_decision_api_v1_aoq_decision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aoq/evaluate": {
      "post": {
        "tags": [
          "AOQ"
        ],
        "summary": "Evaluate Decision",
        "operationId": "evaluate_decision_api_v1_aoq_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aoq/rules": {
      "get": {
        "tags": [
          "AOQ"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_aoq_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  },
                  "type": "array",
                  "title": "Response List Rules Api V1 Aoq Rules Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AOQ"
        ],
        "summary": "Create Rule",
        "operationId": "create_rule_api_v1_aoq_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aoq/decisions/{decision_id}": {
      "get": {
        "tags": [
          "AOQ"
        ],
        "summary": "Get Decision",
        "operationId": "get_decision_api_v1_aoq_decisions__decision_id__get",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aoq/audit/{entity_id}": {
      "get": {
        "tags": [
          "AOQ"
        ],
        "summary": "Get Audit",
        "operationId": "get_audit_api_v1_aoq_audit__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditTrailResponse"
                  },
                  "title": "Response Get Audit Api V1 Aoq Audit  Entity Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AoqFeatures": {
        "properties": {
          "fintech_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Fintech Score"
          },
          "mobility_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Mobility Score"
          },
          "esg_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Esg Score"
          },
          "social_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Social Score"
          }
        },
        "type": "object",
        "required": [
          "fintech_score",
          "mobility_score",
          "esg_score",
          "social_score"
        ],
        "title": "AoqFeatures"
      },
...
      "AuditTrailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "entity_id",
          "payload",
          "signature",
          "created_at"
        ],
        "title": "AuditTrailResponse"
      },
...
      "DecisionRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id"
          },
          "signal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Id"
          },
          "features": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AoqFeatures"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "DecisionRequest"
      },
      "DecisionResponse": {
        "properties": {
          "decision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Id"
          },
          "signal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Signal Id"
          },
          "rule_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rule Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "decision_id",
          "signal_id",
          "rule_id",
          "user_id",
          "score",
          "threshold",
          "decision",
          "rationale",
          "created_at"
        ],
        "title": "DecisionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "weight_fintech": {
            "type": "number",
            "title": "Weight Fintech"
          },
          "weight_mobility": {
            "type": "number",
            "title": "Weight Mobility"
          },
          "weight_esg": {
            "type": "number",
            "title": "Weight Esg"
          },
          "weight_social": {
            "type": "number",
            "title": "Weight Social"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "threshold",
          "weight_fintech",
          "weight_mobility",
          "weight_esg",
          "weight_social",
          "active",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "RuleResponse"
      },
      "RuleSchema": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Name"
          },
          "threshold": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Threshold"
          },
          "weight_fintech": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight Fintech",
            "default": 0.35
          },
          "weight_mobility": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight Mobility",
            "default": 0.25
          },
          "weight_esg": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight Esg",
            "default": 0.25
          },
          "weight_social": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight Social",
            "default": 0.15
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "threshold"
        ],
        "title": "RuleSchema"
      },
...
      "SignalRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id"
          },
          "source": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Source",
            "default": "mobile"
          },
          "features": {
            "$ref": "#/components/schemas/AoqFeatures"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "features"
        ],
        "title": "SignalRequest"
      },
      "SignalResponse": {
        "properties": {
          "signal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Signal Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "signal_id",
          "user_id",
          "source",
          "created_at"
        ],
        "title": "SignalResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}